# openapi.json
{
"openapi": "3.1.0",
"info": {
"title": "NEXUS PM Agent API",
"version": "0.1.0"
},
"paths": {
...
"/config": {
"get": {
"tags": [
"config"
],
"summary": "Get Config",
"description": "404 when no config row exists yet \u2014 'not saved yet' is not a defect;\n422-with-defects is reserved for rejected SAVES.",
"operationId": "get_config_config_get",
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Response Get Config Config Get"
}
}
}
}
}
},
"put": {
"tags": [
"config"
],
"summary": "Put Config",
"operationId": "put_config_config_put",
"requestBody": {
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Body"
}
}
},
"required": true
},
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Response Put Config Config Put"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/overrides": {
"put": {
"tags": [
"config"
],
"summary": "Put Overrides",
"operationId": "put_overrides_projects__project_id__overrides_put",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Body"
}
}
}
},
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Put Overrides Projects  Project Id  Overrides Put"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/config": {
"get": {
"tags": [
"config"
],
"summary": "Get Resolved Config",
"description": "The RESOLVED view (project override first, client default otherwise) so\noverride behavior is visible during testing. Keys that fail resolution\n(required-missing-at-both-levels) surface as their defect text.",
"operationId": "get_resolved_config_projects__project_id__config_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Get Resolved Config Projects  Project Id  Config Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
...
"/status-reports": {
"post": {
"tags": [
"registers"
],
"summary": "Submit Status Report",
"description": "The confirmed-Q7 manual inbox: the row is parsed by Status Tracking on\nthe next monitoring cycle, not here.\n\nScoping: the task must belong to the caller's company, and a member may\nonly report AS THEMSELVES (their linked roster row) \u2014 client_admin can\nfile on anyone's behalf.",
"operationId": "submit_status_report_status_reports_post",
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/StatusReportBody"
}
}
},
"required": true
},
"responses": {
"201": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Response Submit Status Report Status Reports Post"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/status-reports": {
"get": {
"tags": [
"registers"
],
"summary": "List Status Reports",
"operationId": "list_status_reports_projects__project_id__status_reports_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response List Status Reports Projects  Project Id  Status Reports Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/risks": {
"get": {
"tags": [
"registers"
],
"summary": "List Risks",
"operationId": "list_risks_projects__project_id__risks_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
},
{
"name": "status",
"in": "query",
"required": false,
"schema": {
"anyOf": [
{
"type": "string"
},
{
"type": "null"
}
],
"title": "Status"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response List Risks Projects  Project Id  Risks Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/risks/{risk_id}/score": {
"patch": {
"tags": [
"registers"
],
"summary": "Adjust Risk Score",
"description": "Wraps risk_tracking.adjust_score (OQ-6: audited, human actor).",
"operationId": "adjust_risk_score_risks__risk_id__score_patch",
"parameters": [
{
"name": "risk_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Risk Id"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ScoreBody"
}
}
}
},
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Adjust Risk Score Risks  Risk Id  Score Patch"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/blockers": {
"get": {
"tags": [
"registers"
],
"summary": "List Blockers",
"description": "Unowned blockers (assigned_to NULL) sort first \u2014 the UI surfaces them.",
"operationId": "list_blockers_projects__project_id__blockers_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response List Blockers Projects  Project Id  Blockers Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/blockers/{blocker_id}": {
"patch": {
"tags": [
"registers"
],
"summary": "Patch Blocker",
"description": "Wraps blockers.assign_blocker / resolve_blocker (OQ-6: audited).",
"operationId": "patch_blocker_blockers__blocker_id__patch",
"parameters": [
{
"name": "blocker_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Blocker Id"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/BlockerPatch"
}
}
}
},
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Patch Blocker Blockers  Blocker Id  Patch"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/escalation-log": {
"get": {
"tags": [
"registers"
],
"summary": "Escalation Log",
"operationId": "escalation_log_projects__project_id__escalation_log_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response Escalation Log Projects  Project Id  Escalation Log Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/audit-log": {
"get": {
"tags": [
"registers"
],
"summary": "Audit Log",
"operationId": "audit_log_projects__project_id__audit_log_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
},
{
"name": "limit",
"in": "query",
"required": false,
"schema": {
"type": "integer",
"default": 200,
"title": "Limit"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response Audit Log Projects  Project Id  Audit Log Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/projects/{project_id}/artifacts": {
"get": {
"tags": [
"registers"
],
"summary": "Artifacts",
"operationId": "artifacts_projects__project_id__artifacts_get",
"parameters": [
{
"name": "project_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Project Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"type": "object",
"additionalProperties": true
},
"title": "Response Artifacts Projects  Project Id  Artifacts Get"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
}
},
"components": {
"schemas": {
"BlockerPatch": {
"properties": {
"assigned_to": {
"anyOf": [
{
"type": "integer"
},
{
"type": "null"
}
],
"title": "Assigned To"
},
"resolve": {
"type": "boolean",
"title": "Resolve",
"default": false
}
},
"type": "object",
"title": "BlockerPatch"
},
...
"HTTPValidationError": {
"properties": {
"detail": {
"items": {
"$ref": "#/components/schemas/ValidationError"
},
"type": "array",
"title": "Detail"
}
},
"type": "object",
"title": "HTTPValidationError"
},
...
"ScoreBody": {
"properties": {
"severity": {
"type": "integer",
"title": "Severity"
},
"likelihood": {
"type": "integer",
"title": "Likelihood"
}
},
"type": "object",
"required": [
"severity",
"likelihood"
],
"title": "ScoreBody"
},
...
"StatusReportBody": {
"properties": {
"task_id": {
"type": "integer",
"title": "Task Id"
},
"member_id": {
"type": "integer",
"title": "Member Id"
},
"raw_text": {
"type": "string",
"title": "Raw Text"
}
},
"type": "object",
"required": [
"task_id",
"member_id",
"raw_text"
],
"title": "StatusReportBody"
},
...
"ValidationError": {
"properties": {
"loc": {
"items": {
"anyOf": [
{
"type": "string"
},
{
"type": "integer"
}
]
},
"type": "array",
"title": "Location"
},
"msg": {
"type": "string",
"title": "Message"
},
"type": {
"type": "string",
"title": "Error Type"
},
"input": {
"title": "Input"
},
"ctx": {
"type": "object",
"title": "Context"
}
},
"type": "object",
"required": [
"loc",
"msg",
"type"
],
"title": "ValidationError"
}
}
}
}